mdp-deepflowFiles/deepflow/{frontend,backend,docs}/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain-google-genai==2.0.7
langchain-core==0.3.15
langgraph==0.2.38
